requirements.txt
# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.0

# AWS integration
boto3>=1.26.0

# Image processing
Pillow>=8.0.0

# Data handling
numpy>=1.21.0

# Optional: For enhanced progress bars and utilities
# rich>=12.0.0

